feat: move secret references replacement form generating to applying
