+++ pkg/secrets/providers/azure/keyvault/keyvault_test.go
@@ -156,7 +156,7 @@ func TestNewSecretStore(t *testing.T) {
 				cleanup := fake.SetClientIDSecretInEnv()
 				defer cleanup()
 			}
-			_, err := factory.NewSecretStore(tc.spec)
+			_, err := factory.NewSecretStore(&tc.spec)
 			if diff := cmp.Diff(err, tc.expectedErr, EquateErrors()); diff != "" {
 				t.Errorf("\n%s\ngot unexpected error:\n%s", name, diff)
 			}
